backend/requirements.txt
# FastAPI 后端依赖

# Web 框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# 数据验证
pydantic>=2.10.5
pydantic-settings>=2.1.0
email-validator>=2.0.0

# 异步支持
httpx>=0.26.0
aiofiles>=23.2.1

# 限流
slowapi>=0.1.9

# CORS (使用 FastAPI 内置的 CORSMiddleware，无需额外包)

# 数据库
sqlalchemy>=2.0.25
alembic>=1.13.1
aiosqlite>=0.19.0
greenlet>=3.0.0

# 认证和授权
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# 工具库
python-dotenv>=1.0.0
loguru>=0.7.2

# SmartPaper 核心依赖
langchain>=0.3.14
langchain-community>=0.3.14
openai>=1.3.7
zhipuai>=1.0.7
python-docx>=1.0.1
python-pptx>=0.6.21
openpyxl>=3.1.2
pytesseract>=0.3.10
Pillow>=10.1.0
beautifulsoup4>=4.12.2
requests>=2.31.0
pyyaml>=6.0.1
pandas>=2.1.4
arxiv>=2.0.0
tqdm>=4.66.1
markitdown[pdf]
pymupdf
langchain_openai
modelscope

# 开发依赖
pre-commit>=3.6.2
black>=24.2.0
isort>=5.13.2
flake8>=7.0.0
pytest>=7.4.4
pytest-asyncio>=0.23.3
